Ensure that only caches with delete_pattern functionality are supported (#37)
* Ensure that only caches with delete_pattern functionality are supported

* Added a function to delete multiple annotations at once

http://127.0.0.1:8000/api/v1/annotations/annotations/multiple_delete/?id__in=1,2,3

+++ exact/exact/administration/models.py
@@ -44,7 +44,8 @@ class Product(models.Model):
 @receiver([post_save, post_delete], sender=Product)
 def product_changed_handler(sender, instance, **kwargs):
     # delte cached imageset information used in JS
-    for image_set in instance.imagesets.all():
-        cache.delete_pattern(f"*/api/v1/images/image_sets/{image_set.id}/*")
+    if hasattr(cache, "delete_pattern"):
+        for image_set in instance.imagesets.all():
+            cache.delete_pattern(f"*/api/v1/images/image_sets/{image_set.id}/*")
 
 m2m_changed.connect(product_changed_handler, sender=Product.imagesets.through)
